# boot/swagger-rest-json/swagger/rest_json.yaml
swagger: '2.0'
info:
  description: REST JSON jsample
  version: '1.0'
  title: REST JSON jsample Api Documentation
host: 'localhost:8080'
basePath: /
tags:
  - name: application-controller
    description: the application controller API
paths:
  /app/getApps:
    get:
      tags:
        - application-controller
      summary: Get list of applications
      operationId: getApps      
      produces:
        - application/json
      responses:
        '200':
          description: application list retrival is successful
          schema:
            type: array
            items:
              $ref: '#/definitions/Application'
  /app/addApp:
    post:
      tags:
        - application-controller
      summary: Add new application
      operationId: addAdd
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: Application
          description: application model
          required: true
          schema:
            $ref: '#/definitions/Application'
      responses:
        '200':
          description: application addition is successful
          schema:
            $ref: '#/definitions/Application'
  /app/updateApp/{id}:
    put:
      tags:
        - application-controller
      summary: Update existing application
      operationId: updateApp
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: application
          description: Application Model
          required: true
          schema:
            $ref: '#/definitions/Application'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32	
      responses:
        '200':
          description: Application update is successful
          schema:
            $ref: '#/definitions/Application'
  /app/deleteApp/{id}:
    delete:
      tags:
        - application-controller
      summary: Delete application 
      operationId: deleteApp
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Application delete is successful
          schema:
            type: string
definitions:
  Application:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: id of the application
      name:
        type: string
        description: name of the application
      status:
        type: string
        enum:
          - stopped
          - running
        description: status of the application